backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
torch>=2.0.0
torchvision>=0.15.0
pillow>=10.0.0
numpy>=1.26.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
groq>=0.9.0
opencv-python-headless>=4.8.0
python-dotenv>=1.0.0
pydantic>=2.0.0
